# api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://health-tracker.grayflower-c85606c1.westeurope.azurecontainerapps.io
  description: Generated server url
paths:
  /api/v1/workouts-stats/{id}:
    get:
      tags:
      - workout-stat-controller
      summary: Get a workout stat by id
      operationId: getById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Workout stat retrieved successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/WorkoutStatResource'
        "404":
          description: Workout stat not found
          content:
            text/plain: {}
    put:
      tags:
      - workout-stat-controller
      summary: Update a workout stat by id
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutStatResource'
        required: true
      responses:
        "409":
          description: Workout stat is not changed
          content:
            text/plain: {}
        "200":
          description: Workout stat updated successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/WorkoutStatResource'
        "404":
          description: Workout stat not found
          content:
            text/plain: {}
    delete:
      tags:
      - workout-stat-controller
      summary: Delete a workout stat by id
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Workout stat deleted successfully
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Workout stat not found
          content:
            text/plain: {}
...
  /api/v1/sleep-stats/{id}:
    get:
      tags:
      - sleep-stat-controller
      summary: Get a sleep stat by id
      operationId: getById_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Sleep stat not found
          content:
            text/plain: {}
        "200":
          description: Sleep stat retrieved successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SleepStatResource'
    put:
      tags:
      - sleep-stat-controller
      summary: Update a sleep stat by id
      operationId: update_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SleepStatResource'
        required: true
      responses:
        "409":
          description: Sleep stat is not changed
          content:
            text/plain: {}
        "200":
          description: Sleep stat updated successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SleepStatResource'
        "404":
          description: Sleep stat not found
          content:
            text/plain: {}
    delete:
      tags:
      - sleep-stat-controller
      summary: Delete a sleep stat by id
      operationId: delete_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Sleep stat not found
          content:
            text/plain: {}
        "204":
          description: Sleep stat deleted successfully
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/meals/{id}:
    get:
      tags:
      - meal-controller
      summary: Get a meal by id
      operationId: getById_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Meal not found
          content:
            text/plain: {}
        "200":
          description: Meal retrieved successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MealResource'
    put:
      tags:
      - meal-controller
      summary: Update a meal by id
      operationId: update_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MealResource'
        required: true
      responses:
        "409":
          description: Meal is not changed
          content:
            text/plain: {}
        "404":
          description: Meal not found
          content:
            text/plain: {}
        "200":
          description: Meal updated successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MealResource'
    delete:
      tags:
      - meal-controller
      summary: Delete a meal by id
      operationId: delete_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Meal not found
          content:
            text/plain: {}
        "204":
          description: Meal deleted successfully
          content:
            '*/*':
              schema:
                type: object
  /api/v1/meals-stats/{id}:
    get:
      tags:
      - meal-stat-controller
      summary: Get a meal stat by id
      operationId: getById_5
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Meal stat retrieved successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MealStatResource'
        "404":
          description: Meal stat not found
          content:
            text/plain: {}
    put:
      tags:
      - meal-stat-controller
      summary: Update a meal stat by id
      operationId: update_5
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MealStatResource'
        required: true
      responses:
        "200":
          description: Meal stat updated successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MealStatResource'
        "409":
          description: Meal stat is not changed
          content:
            text/plain: {}
        "404":
          description: Meal stat not found
          content:
            text/plain: {}
    delete:
      tags:
      - meal-stat-controller
      summary: Delete a meal stat by id
      operationId: delete_5
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Meal stat deleted successfully
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Meal stat not found
          content:
            text/plain: {}
  /api/v1/workouts-stats:
    get:
      tags:
      - workout-stat-controller
      summary: Get all workout stats
      operationId: getAll
      responses:
        "200":
          description: All workout stats retrieved successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkoutStatResource'
    post:
      tags:
      - workout-stat-controller
      summary: Create a workout stat
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutStatResource'
        required: true
      responses:
        "404":
          description: Foreign key constraint violation
          content:
            text/plain: {}
        "201":
          description: Workout stat created successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/WorkoutStatResource'
  /api/v1/sleep-stats:
    get:
      tags:
      - sleep-stat-controller
      summary: Get all sleep stats
      operationId: getAll_2
      responses:
        "200":
          description: All sleep stats retrieved successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SleepStatResource'
    post:
      tags:
      - sleep-stat-controller
      summary: Create a new sleep stat
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SleepStatResource'
        required: true
      responses:
        "201":
          description: Sleep stat created successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SleepStatResource'
        "404":
          description: Foreign key constraint violation
          content:
            text/plain: {}
...
  /api/v1/meals-stats:
    get:
      tags:
      - meal-stat-controller
      summary: Get all meal stats
      operationId: getAll_5
      responses:
        "200":
          description: All meal stats retrieved successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MealStatResource'
    post:
      tags:
      - meal-stat-controller
      summary: Create a new meal stat
      operationId: create_3
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MealStatResource'
        required: true
      responses:
        "404":
          description: Foreign key constraint violation
          content:
            text/plain: {}
        "201":
          description: Meal stat created successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MealStatResource'
...
components:
  schemas:
    WorkoutStatResource:
      required:
      - burned_calories
      - date
      - duration
      - user_id
      type: object
      properties:
        id:
          type: integer
          format: int64
        duration:
          minimum: 1
          type: integer
          format: int32
        date:
          type: string
        user_id:
          type: integer
          format: int64
        burned_calories:
          minimum: 1
          type: integer
          format: int32
        workout_bio:
          type: string
...
    SleepStatResource:
      required:
      - date
      - duration
      - quality_status
      - user_id
      type: object
      properties:
        id:
          type: integer
          format: int64
        duration:
          minimum: 1
          type: integer
          format: int32
        date:
          type: string
        user_id:
          type: integer
          format: int64
        quality_status:
          type: string
          enum:
          - GOOD
          - BAD
          - NEUTRAL
...
    MealResource:
      required:
      - calories
      - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          maxLength: 64
          minLength: 2
          type: string
        calories:
          minimum: 0
          type: integer
          format: int32
    MealStatResource:
      required:
      - date
      - meal_id
      - user_id
      type: object
      properties:
        id:
          type: integer
          format: int64
        date:
          type: string
        meal_id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
...